+++ src/main/evaluation/index.scala
@@ -5,7 +5,7 @@ import translation.structure.*
 import evaluation.expression.evaluateBasicExpr
 import evaluation.config.*
 
-def evaluateIndex(index: Index): Int = index match {
+def evaluateIndex(index: Index): Double = index match {
     case ExprIndex(value) => ExprIndexCost + evaluateBasicExpr(value)
     case StartSubIndex(value) => StartSubIndexCost + evaluateBasicExpr(value)
     case UpdateIndex(left, right) =>
